requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.0
numpy==1.26.4
yfinance==0.2.40
beanie==1.26.0
motor==3.1.2
pymongo==4.3.3
mangum==0.17.0
python-dotenv==1.0.1
pycryptodome==3.23.0
finance-datareader==0.9.96
fredapi==0.5.0
OpenDartReader==0.1.1
celery==5.4.0
redis==5.0.4
websocket-client==1.8.0
fastparquet==2024.11.0
holidays==0.87
python-telegram-bot==21.3